enhanced-puter-api/requirements.txt
# Enhanced Puter API Wrapper - Dependencies
# Flask Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Werkzeug==3.0.1

# Puter SDK and API clients
putergenai>=0.1.5
requests>=2.31.0
aiohttp>=3.9.0

# Data handling and validation
pydantic>=2.5.0
python-multipart>=0.0.6
marshmallow>=3.20.0

# Production server
gunicorn>=21.2.0
gevent>=23.9.1

# Environment and configuration
python-dotenv>=1.0.0
PyYAML>=6.0.1

# Optional Redis support (comment out if not using Redis)
# redis>=5.0.0
# flask-caching>=2.1.0

# Development and testing
pytest>=7.4.0
pytest-flask>=1.3.0

# Logging and monitoring
structlog>=23.2.0
